chore: extract node functions to its own file
